+++ Application/Models/Medium.cs
@@ -11,7 +11,7 @@ public partial class Medium
 
     public string? Status { get; set; }
 
-    public byte Count { get; set; }
+    public int Count { get; set; }
 
     public string? Poster { get; set; }
